Address CodeRabbit UUID fixture feedback (DEBT-400 PR 3b)
Validate and apply the actionable CodeRabbit feedback on PR #372.\n\n- Move question-page browser UUID fixture constants into a pure shared fixture module and re-export them from the existing helper to avoid importing mock setup side effects.\n- Replace the remaining browser DTO choice/question ID placeholders in practice-view and exam-review viewport fixtures with generated UUIDs.\n- Tie the DEBT-400 PR 3b doc status to PR #372.\n\nNo fixture vi.hoisted() blocks were added; the browser spy/mockResolvedValue pattern still uses normal-scope UUIDs.
