+++ app/(app)/app/questions/[slug]/use-question-page-controller-test-helpers.tsx
@@ -8,7 +8,19 @@ import * as questionViewController from '@/src/adapters/controllers/question-vie
 import type { GetBookmarksOutput } from '@/src/application/ports/bookmarks';
 import { ok } from '@/tests/test-helpers/ok';
 import { installReportClientErrorMocks } from '@/tests/test-helpers/report-client-error-mocks';
+import { QUESTION_PAGE_CHOICE_1_ID } from './question-page-controller.browser.fixtures';
 import { useQuestionPageController } from './use-question-page-controller';
+
+export {
+  getQuestionPageQuestionIdForSlug,
+  QUESTION_PAGE_ATTEMPT_1_ID,
+  QUESTION_PAGE_ATTEMPT_2_ID,
+  QUESTION_PAGE_ATTEMPT_3_ID,
+  QUESTION_PAGE_CHOICE_1_ID,
+  QUESTION_PAGE_CHOICE_2_ID,
+  QUESTION_PAGE_QUESTION_1_ID,
+  QUESTION_PAGE_QUESTION_2_ID,
+} from './question-page-controller.browser.fixtures';
 
 vi.mock('@/src/adapters/controllers/question-view-controller', { spy: true });
 vi.mock('@/src/adapters/controllers/question-controller', { spy: true });
@@ -33,20 +45,6 @@ export const reportClientErrorSpy = vi.mocked(
 );
 
 installReportClientErrorMocks(reportClientError);
-
-export const QUESTION_PAGE_QUESTION_1_ID = crypto.randomUUID();
-export const QUESTION_PAGE_QUESTION_2_ID = crypto.randomUUID();
-export const QUESTION_PAGE_CHOICE_1_ID = crypto.randomUUID();
-export const QUESTION_PAGE_CHOICE_2_ID = crypto.randomUUID();
-export const QUESTION_PAGE_ATTEMPT_1_ID = crypto.randomUUID();
-export const QUESTION_PAGE_ATTEMPT_2_ID = crypto.randomUUID();
-export const QUESTION_PAGE_ATTEMPT_3_ID = crypto.randomUUID();
-
-export function getQuestionPageQuestionIdForSlug(slug: string) {
-  return slug === 'q-2'
-    ? QUESTION_PAGE_QUESTION_2_ID
-    : QUESTION_PAGE_QUESTION_1_ID;
-}
 
 const emptyBookmarksResult: { ok: true; data: GetBookmarksOutput } = ok({
   rows: [],
