+++ app/(app)/app/questions/[slug]/use-question-page-session-navigation.browser.spec.tsx
@@ -6,10 +6,11 @@ import type { QuestionOrigin } from '@/lib/routes';
 import * as practiceController from '@/src/adapters/controllers/practice-controller';
 import { ok } from '@/tests/test-helpers/ok';
 import { installReportClientErrorMocks } from '@/tests/test-helpers/report-client-error-mocks';
+import {
+  QUESTION_PAGE_QUESTION_1_ID,
+  QUESTION_PAGE_QUESTION_2_ID,
+} from './question-page-controller.browser.fixtures';
 import { useQuestionPageSessionNavigation } from './use-question-page-session-navigation';
-
-const fixtureQuestion1Id = crypto.randomUUID();
-const fixtureQuestion2Id = crypto.randomUUID();
 
 vi.mock('@/src/adapters/controllers/practice-controller', { spy: true });
 vi.mock('@/lib/report-client-error', { spy: true });
@@ -86,7 +87,7 @@ describe('useQuestionPageSessionNavigation (browser)', () => {
         rows: [
           {
             isAvailable: true,
-            questionId: fixtureQuestion1Id,
+            questionId: QUESTION_PAGE_QUESTION_1_ID,
             slug: 'q-1',
             stemMd: 'Stem',
             difficulty: 'easy',
@@ -98,7 +99,7 @@ describe('useQuestionPageSessionNavigation (browser)', () => {
           },
           {
             isAvailable: true,
-            questionId: fixtureQuestion2Id,
+            questionId: QUESTION_PAGE_QUESTION_2_ID,
             slug: 'q-2',
             stemMd: 'Stem 2',
             difficulty: 'easy',
@@ -141,7 +142,7 @@ describe('useQuestionPageSessionNavigation (browser)', () => {
         rows: [
           {
             isAvailable: true,
-            questionId: fixtureQuestion1Id,
+            questionId: QUESTION_PAGE_QUESTION_1_ID,
             slug: 'q-1',
             stemMd: 'Stem',
             difficulty: 'easy',
@@ -153,7 +154,7 @@ describe('useQuestionPageSessionNavigation (browser)', () => {
           },
           {
             isAvailable: true,
-            questionId: fixtureQuestion2Id,
+            questionId: QUESTION_PAGE_QUESTION_2_ID,
             slug: 'q-2',
             stemMd: 'Stem 2',
             difficulty: 'easy',
